refactor: move shared_specs to shotgun_web and simplify compact layouts
- Move src/shotgun/shared_specs/ to src/shotgun/shotgun_web/shared_specs/
  for better code organization (module tightly coupled with shotgun_web)
- Move test/unit/shared_specs/ to test/unit/shotgun_web/shared_specs/
- Update all imports to use new module path
- Simplify _apply_compact_layout() in TUI dialogs by using screen-level
  CSS classes instead of element-level classes (reduces code duplication)
