+++ src/shotgun/shotgun_web/shared_specs/__init__.py
@@ -4,20 +4,20 @@ This module provides utilities for scanning and hashing files in the
 .shotgun/ directory for upload to the shared specs API.
 """
 
-from shotgun.shared_specs.file_scanner import (
+from shotgun.shotgun_web.shared_specs.file_scanner import (
     get_shotgun_directory,
     scan_shotgun_directory,
 )
-from shotgun.shared_specs.hasher import (
+from shotgun.shotgun_web.shared_specs.hasher import (
     calculate_sha256,
     calculate_sha256_with_size,
 )
-from shotgun.shared_specs.models import (
+from shotgun.shotgun_web.shared_specs.models import (
     UploadProgress,
     UploadResult,
 )
-from shotgun.shared_specs.upload_pipeline import run_upload_pipeline
-from shotgun.shared_specs.utils import UploadPhase, format_bytes
+from shotgun.shotgun_web.shared_specs.upload_pipeline import run_upload_pipeline
+from shotgun.shotgun_web.shared_specs.utils import UploadPhase, format_bytes
 
 __all__ = [
     "UploadPhase",
