+++ src/shotgun/shotgun_web/shared_specs/file_scanner.py
