+++ src/shotgun/shotgun_web/shared_specs/hasher.py
