+++ src/shotgun/shotgun_web/shared_specs/utils.py
